requirements.txt
python-dotenv
requests
mysql-connector-python
flask>=3.0,<4
pytest>=8.0,<9
